+++ src/controllers/CardController.ts
@@ -1,32 +1,41 @@
-import { Request, Response } from 'express'
+import {  Response } from 'express'
+import { v4 as uuid } from 'uuid'
+
+import { IUserRequest } from '../types/Auth.types'
 
 import { respondCreated, respondNotFound, respondOk, respondServerError } from '../utils/responses'
 
 import Card from '../models/Card'
 
-export const getCards = async (req: Request, res: Response) => {
+export const getCards = async (req: IUserRequest, res: Response) => {
     try {
-        const cards = await Card.query()
+        const cards = await Card.query().where('user_id', '=', req.user.id)
         return respondOk(req, res, { cards })
     } catch (error: any) {
         return respondServerError(req, res, null, 'Something went wrong processing your request', 500, error.message)
     }
 }
 
-export const getSingleCard = async (req: Request, res: Response) => {
+export const getSingleCard = async (req: IUserRequest, res: Response) => {
     try {
-        const card = await Card.query().findById(req.params.id)
+        const card = await Card.query().findById(req.params.id).where('user_id', '=', req.user.id)
         return respondOk(req, res, { card })
     } catch (error: any) {
         return respondServerError(req, res, null, 'Something went wrong processing your request', 500, error.message)
     }
 }
 
-export const createSingleCard = async (req: Request, res: Response) => {
+export const createSingleCard = async (req: IUserRequest, res: Response) => {
     try {
         const now = new Date().toISOString()
-        const body = { ...req.body, isDefault: Boolean(req.body.isDefault), createdOn: now, updatedOn: now }
-        delete body.id
+        const body = {
+            ...req.body,
+            isDefault: Boolean(req.body.isDefault),
+            createdOn: now,
+            updatedOn: now,
+            userId: req.user.id,
+            id: uuid(),
+        }
         const card = await Card.query().insertAndFetch(body)
         return respondOk(req, res, { card })
     } catch (error: any) {
@@ -34,9 +43,9 @@ export const createSingleCard = async (req: Request, res: Response) => {
     }
 }
 
-export const updateSingleCard = async (req: Request, res: Response) => {
+export const updateSingleCard = async (req: IUserRequest, res: Response) => {
     try {
-        const stagedCard = await Card.query().findById(req.params.id)
+        const stagedCard = await Card.query().findById(req.params.id).where('user_id', '=', req.user.id)
     
         if (!stagedCard) {
             return respondNotFound(req, res, null, `No card with id "${req.params.id}" found.`)
@@ -66,9 +75,9 @@ export const updateSingleCard = async (req: Request, res: Response) => {
     }
 }
 
-export const deleteSingleCard = async (req: Request, res: Response) => {
+export const deleteSingleCard = async (req: IUserRequest, res: Response) => {
     try {
-        await Card.query().deleteById(req.params.id)
+        await Card.query().deleteById(req.params.id).where('user_id', '=', req.user.id)
 
         return respondCreated(req, res, null, 'Card deleted successfully')
     } catch (error: any) {
@@ -76,9 +85,9 @@ export const deleteSingleCard = async (req: Request, res: Response) => {
     }
 }
 
-export const setActiveCard = async (req: Request, res: Response) => {
+export const setActiveCard = async (req: IUserRequest, res: Response) => {
     try {
-        const actives = await Card.query().where('is_default', true)
+        const actives = await Card.query().where('is_default', true).where('user_id', '=', req.user.id)
 
         for (const activeCard of actives) {
             activeCard.$query().patch({ isDefault: false })
